refactor: move and rename SchoolPicker
