+++ src/components/AppComponents/LayoutPreviewControls/LayoutPreviewControls.test.tsx
@@ -1,12 +1,13 @@
-import { ToastProvider } from "../../context/Toast";
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
+import LayoutPreviewControls from "./LayoutPreviewControls";
 
-import PreviewControls from "./PreviewControls";
+import { ToastProvider } from "@/context/Toast";
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const useRouter = jest.spyOn(require("next/router"), "useRouter");
 
-describe("PreviewControls", () => {
+describe("LayoutPreviewControls", () => {
   beforeEach(() => {
     jest.resetAllMocks();
     jest.clearAllMocks();
@@ -20,7 +21,7 @@ describe("PreviewControls", () => {
 
     const { getByRole } = renderWithTheme(
       <ToastProvider>
-        <PreviewControls />
+        <LayoutPreviewControls />
       </ToastProvider>,
     );
