+++ src/components/TeacherComponents/ResourcePageSchoolPicker/formatSchoolName.test.tsx
