+++ src/components/TeacherComponents/ResourcePageSchoolPicker/formatSchoolName.tsx
