+++ src/components/TeacherComponents/ResourcePageSchoolPicker/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ResourcePageSchoolPicker";
